minor daemon.py refactoring

+++ fuglu/src/fuglu/daemon.py
@@ -27,12 +27,15 @@ class DaemonStuff(object):
     def __init__(self, pidfilename):
         self.pidfile = pidfilename
 
+
     def delpid(self):
         """Delete the pid file"""
         try:
-            os.remove(self.pidfile)
-        except:
+            if os.path.exists(self.pidfile):
+                os.remove(self.pidfile)
+        except Exception:
             pass
+
 
     def createDaemon(self):
         """Detach a process from the controlling terminal and run it in the
@@ -45,14 +48,14 @@ class DaemonStuff(object):
         except OSError as e:
             raise Exception("%s [%d]" % (e.strerror, e.errno))
 
-        if (pid == 0):
+        if pid == 0:
             os.setsid()
             try:
                 pid = os.fork()    # Fork a second child.
             except OSError as e:
                 raise Exception("%s [%d]" % (e.strerror, e.errno))
 
-            if (pid == 0):    # The second child.
+            if pid == 0:    # The second child.
                 os.chdir('/')
                 os.umask(0)
             else:
@@ -64,7 +67,7 @@ class DaemonStuff(object):
 
         import resource        # Resource usage information.
         maxfd = resource.getrlimit(resource.RLIMIT_NOFILE)[1]
-        if (maxfd == resource.RLIM_INFINITY):
+        if maxfd == resource.RLIM_INFINITY:
             maxfd = 1024
 
         # Iterate through and close all file descriptors.
@@ -85,7 +88,8 @@ class DaemonStuff(object):
         pidfd = os.open(self.pidfile, os.O_WRONLY | os.O_CREAT, 0o644)
         os.write(pidfd, "%s\n" % pid)
         os.close(pidfd)
-        return(0)
+        return 0
+
 
     def drop_privs(self, username='nobody', groupname='nobody', keep_supplemental_groups=True):
         """Drop privileges of the current process to specified unprivileged user and group. If keep_supplemental_groups is True,
@@ -94,7 +98,7 @@ class DaemonStuff(object):
         try:
             running_uid = pwd.getpwnam(username).pw_uid
             running_gid = grp.getgrnam(groupname).gr_gid
-        except:
+        except Exception:
             raise Exception('Can not drop privileges, user %s or group %s does not exist' % (
                 username, groupname))
         new_umask = 0o077
@@ -104,6 +108,7 @@ class DaemonStuff(object):
         if keep_supplemental_groups:
             os.setgroups(self._get_group_ids(username))
         os.setuid(running_uid)
+
 
     def _get_group_ids(self, username):
         """Return a list of group ids the user belongs to"""
